+++ StateR.sln
@@ -13,6 +13,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{372F8647-9
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StateR.Tests", "test\StateR.Tests\StateR.Tests.csproj", "{3276327A-63D9-4BEA-BBB8-D3AFF841CC67}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StateR.Experiments", "src\StateR.Experiments\StateR.Experiments.csproj", "{0EB20F7F-8AA1-48E2-9489-6975566469AF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StateR.Blazor.Experiments", "src\StateR.Blazor.Experiments\StateR.Blazor.Experiments.csproj", "{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -59,6 +63,30 @@ Global
 		{3276327A-63D9-4BEA-BBB8-D3AFF841CC67}.Release|x64.Build.0 = Release|Any CPU
 		{3276327A-63D9-4BEA-BBB8-D3AFF841CC67}.Release|x86.ActiveCfg = Release|Any CPU
 		{3276327A-63D9-4BEA-BBB8-D3AFF841CC67}.Release|x86.Build.0 = Release|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Debug|x64.Build.0 = Debug|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Debug|x86.Build.0 = Debug|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Release|x64.ActiveCfg = Release|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Release|x64.Build.0 = Release|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Release|x86.ActiveCfg = Release|Any CPU
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF}.Release|x86.Build.0 = Release|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Debug|x64.Build.0 = Debug|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Debug|x86.Build.0 = Debug|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Release|x64.ActiveCfg = Release|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Release|x64.Build.0 = Release|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Release|x86.ActiveCfg = Release|Any CPU
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -67,6 +95,8 @@ Global
 		{1B6B12F7-0851-4B75-89AA-BEFFF6CF04BB} = {F0F6A2CA-0972-43BD-B777-B5656DFE20C3}
 		{EA4C22B6-78A9-481A-85BF-62FF1EA70F1B} = {F0F6A2CA-0972-43BD-B777-B5656DFE20C3}
 		{3276327A-63D9-4BEA-BBB8-D3AFF841CC67} = {372F8647-9F67-4176-B0C1-B0E5CAD87C9E}
+		{0EB20F7F-8AA1-48E2-9489-6975566469AF} = {F0F6A2CA-0972-43BD-B777-B5656DFE20C3}
+		{5B16AF66-5F00-4E5F-AAEF-DE2CF549E21D} = {F0F6A2CA-0972-43BD-B777-B5656DFE20C3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6ADD1933-C449-475E-9409-AD333C1C48A0}
