+++ src/StateR.Blazor.Experiments/compilerconfig.json
@@ -1,0 +1,6 @@
+﻿[
+  {
+    "outputFile": "wwwroot/css/StateR.Blazor.css",
+    "inputFile": "Components/ErrorToastStack.razor.scss"
+  }
+]
