+++ src/StateR.Blazor.Experiments/compilerconfig.json.defaults
