--- src/StateR.Blazor/compilerconfig.json
@@ -1,6 +1,0 @@
-﻿[
-    {
-        "outputFile": "wwwroot/css/StateR.Blazor.css",
-        "inputFile": "Errors/ToastStack.razor.scss"
-    }
-]
