+++ src/StateR.Experiments/AsyncLogic/AsyncOperation.cs
