+++ src/StateR.Experiments/AsyncLogic/AsyncOperationStatus.cs
